requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
alembic==1.17.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
baml-py==0.211.0
    # via -r requirements.in
bcrypt==4.2.0
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via uvicorn
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via pydantic
fastapi==0.115.0
    # via -r requirements.in
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
iniconfig==2.1.0
    # via pytest
loguru==0.7.2
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
mypy==1.18.1
    # via -r requirements.in
mypy-extensions==1.1.0
    # via mypy
packaging==25.0
    # via pytest
pathspec==0.12.1
    # via mypy
pluggy==1.6.0
    # via pytest
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.1
    # via -r requirements.in
pytest==8.3.3
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.24.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-jose==3.3.0
    # via -r requirements.in
rsa==4.9.1
    # via python-jose
ruff==0.13.0
    # via -r requirements.in
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
socksio==1.0.0
    # via httpx
sqlalchemy==2.0.35
    # via
    #   -r requirements.in
    #   alembic
starlette==0.38.6
    # via fastapi
ty==0.0.1a20
    # via -r requirements.in
types-pyasn1==0.6.0.20250914
    # via types-python-jose
types-python-jose==3.5.0.20250531
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uv==0.8.18
    # via -r requirements.in
uvicorn==0.30.6
    # via -r requirements.in
